drivers/amd/atombios: add complete legacy modesetting support
Combine the overlapping AtomBIOS display changes into a conflict-free
revision. Add DP link training and eDP power sequencing, VGA/CRT DAC
support, the complete legacy GPU PCI ID table, atomfirmware v2 table
indices, and the R600/R700 display initialization needed by later fixes.

The original revisions could not be rebased independently after their
shared modesetting code was reorganized, so keep the tested resolved tree
as one change.

+++ src/include/framebuffer_info.h
@@ -12,7 +12,7 @@ struct fb_info;
 struct fb_info *
 fb_add_framebuffer_info_ex(const struct lb_framebuffer *fb);
 
-struct fb_info *fb_add_framebuffer_info(uintptr_t fb_addr, uint32_t x_resolution,
+struct fb_info *fb_add_framebuffer_info(uint64_t fb_addr, uint32_t x_resolution,
 					uint32_t y_resolution, uint32_t bytes_per_line,
 					uint8_t bits_per_pixel);
 int fb_add_framebuffer_info_simple(uintptr_t fb_addr, uint32_t x_res, uint32_t y_res,
